Split `features.ts` up.
