+++ src/main.ts
@@ -7,7 +7,8 @@ import {restore} from './features/cache'
 import {setUpDependencies} from './dependencies'
 import {setUpGUComponents} from './gu'
 import {setUpMandrel} from './mandrel'
-import {checkForUpdates, setUpNativeImageMusl} from './features'
+import {checkForUpdates} from './features/check-for-updates'
+import {setUpNativeImageMusl} from './features/musl'
 import {setUpWindowsEnvironment} from './msvc'
 import {setUpNativeImageBuildReports} from './features/reports'
